+++ mcm4csharp/v1/Data/Conversations/Reply.cs
@@ -4,13 +4,13 @@ using System.Text.Json.Serialization;
 namespace mcm4csharp.v1.Data.Conversations {
 	public struct Reply {
 		[JsonPropertyName ("message_id")]
-		public uint MessageId { get; set; }
+		public ulong MessageId { get; set; }
 
 		[JsonPropertyName ("message_date")]
-		public uint MessageDate { get; set; }
+		public ulong MessageDate { get; set; }
 
 		[JsonPropertyName ("author_id")]
-		public uint AuthorId { get; set; }
+		public ulong AuthorId { get; set; }
 
 		[JsonPropertyName ("message")]
 		public string Message { get; set; }
