+++ mcm4csharp/v1/Data/Resources/Download.cs
@@ -4,16 +4,16 @@ using System.Text.Json.Serialization;
 namespace mcm4csharp.v1.Data.Resources {
 	public struct Download {
 		[JsonPropertyName ("download_id")]
-		public uint DownloadId { get; set; }
+		public ulong DownloadId { get; set; }
 
 		[JsonPropertyName ("version_id")]
-		public uint VersionId { get; set; }
+		public ulong VersionId { get; set; }
 
 		[JsonPropertyName ("downloader_id")]
-		public uint DownloaderId { get; set; }
+		public ulong DownloaderId { get; set; }
 
 		[JsonPropertyName ("download_date")]
-		public uint DownloadDate { get; set; }
+		public ulong DownloadDate { get; set; }
 	}
 }
